Added boost build helpers for MSVC
This adds a boost_build project, which is the owner of the custom_command for
running the boost build (b2).
CppProject :: add_boost_lib_dependency() can be used to easily add a library
dependency.  As usual, if a shared library dependency is added, a runtime
dependency is created as well.