'compress' build.ninja with ninja variables
Each project's projectDir and builtDir are defined as variables.
During task emit, these prefixes are tested for and replaced
with a variable reference.
The tests are imperfect (only current project's vars are checked)
but it still provides a decent amount of savings.

build.xlat_path() calls have not been propagated to all toolchains'
emit() implementations yet.